correcion orden compra

+++ database/migrations/2023_06_08_201309_create_orden_compras_table.php
@@ -18,9 +18,9 @@ return new class extends Migration
             $table->foreignId('proveedor_id')->constrained('proveedors');
             $table->foreignId('condiciones_pagos_id')->constrained('condiciones_pagos');
             $table->string('descripcion');
-            $table->bigInteger('neto');
-            $table->bigInteger('iva');
-            $table->bigInteger('total');
+            $table->decimal('neto', 10, 2);
+            $table->decimal('iva', 10, 2);
+            $table->decimal('total', 10, 2);
             $table->enum('estado',['pendiente', 'parcial', 'completado']);
             $table->timestamps();
         });
